modified Chat model, changed getchat.route to getChat.routes
