+++ server/models/Chats.model.js
@@ -1,6 +1,7 @@
 import { DataTypes, Model } from "sequelize";
 import util from "util";
 import { db } from "../config/db.js";
+import User from "./User.model.js";
 
 export default class Chats extends Model {
   [util.inspect.custom]() {
@@ -12,16 +13,23 @@ Chats.init(
   {
     chatId: {
       type: DataTypes.INTEGER,
-      autoIncrement: true,
       primaryKey: true,
+      autoIncrement: true,
+
     },
     senderId: {
-      type: DataTypes.INTEGER,
-      allowNull: true,
+      type: DataTypes.ARRAY(DataTypes.INTEGER),
+      references: {
+        model: User,
+        key: 'userId'
+      }
     },
     receiverId: {
-      type: DataTypes.INTEGER,
-      allowNull: true,
+      type: DataTypes.ARRAY(DataTypes.INTEGER), // sequelize can support ARRAY datatype with postgreSQL
+      references: {
+        model: User,
+        key: 'userId'
+      }
     },
   },
   {
